requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
gunicorn==20.1.0